fix(vote_processor): addressing Ashwin's comments + fixing lint errors.

+++ program/tests/accounting.rs
@@ -985,7 +985,7 @@ async fn test_process_notarization_vote_basic() {
         timestamp,
         i64::from(vote_state.latest_timestamp().timestamp)
     );
-    assert_eq!(slot, u64::from(vote_state.latest_notarized_slot()));
+    assert_eq!(slot, vote_state.latest_notarized_slot());
     assert_eq!(&block_id, vote_state.latest_notarized_block_id());
     assert_eq!(0, vote_state.latest_finalized_slot());
     assert_eq!(&Hash::default(), vote_state.latest_finalized_block_id());
@@ -1070,7 +1070,7 @@ async fn test_process_finalization_vote_basic() {
     );
     assert_eq!(0, vote_state.latest_notarized_slot());
     assert_eq!(&Hash::default(), vote_state.latest_notarized_block_id());
-    assert_eq!(slot, u64::from(vote_state.latest_finalized_slot()));
+    assert_eq!(slot, vote_state.latest_finalized_slot());
     assert_eq!(&block_id, vote_state.latest_finalized_block_id());
     assert_eq!(0, vote_state.latest_skip_start_slot());
     assert_eq!(0, vote_state.latest_skip_end_slot());
@@ -1147,9 +1147,9 @@ async fn test_process_skip_vote_basic() {
         timestamp,
         i64::from(vote_state.latest_timestamp().timestamp)
     );
-    assert_eq!(0, u64::from(vote_state.latest_notarized_slot()));
+    assert_eq!(0, vote_state.latest_notarized_slot());
     assert_eq!(&Hash::default(), vote_state.latest_notarized_block_id());
-    assert_eq!(0, u64::from(vote_state.latest_finalized_slot()));
+    assert_eq!(0, vote_state.latest_finalized_slot());
     assert_eq!(&Hash::default(), vote_state.latest_finalized_block_id());
     assert_eq!(start_slot, vote_state.latest_skip_start_slot());
     assert_eq!(end_slot, vote_state.latest_skip_end_slot());
